+++ Compiler/script/cs_parser.cpp
@@ -300,7 +300,7 @@ int cc_tokenize(const char*inpl, ccInternalList*targ, ccCompiledScript*scrip) {
     fmem_close(iii);
     targ->write_meta(SMETA_END,0);
     // clear any temporary tpyes set
-    for (int ii = 0; ii < sym.entries.size(); ii++) {
+    for (int ii = 0; (size_t)ii < sym.entries.size(); ii++) {
         if (sym.entries[ii].stype == SYM_TEMPORARYTYPE)
             sym.entries[ii].stype = 0;
     }
@@ -326,7 +326,7 @@ void free_pointer(int spOffset, int zeroCmd, int arraySym, ccCompiledScript *scr
 void free_pointers_from_struct(int structVarSym, ccCompiledScript *scrip) {
     int structType = sym.entries[structVarSym].vartype;
 
-    for (int dd = 0; dd < sym.entries.size(); dd++) {
+    for (int dd = 0; (size_t)dd < sym.entries.size(); dd++) {
         if ((sym.entries[dd].stype == SYM_STRUCTMEMBER) &&
             (sym.entries[dd].extends == structType) &&
             ((sym.entries[dd].flags & SFLG_IMPORTED) == 0) &&
@@ -363,7 +363,7 @@ int remove_locals(int from_level, int just_count, ccCompiledScript *scrip) {
     if (from_level == 0)
         zeroPtrCmd = SCMD_MEMZEROPTRND;
 
-    for (cc=0;cc<sym.entries.size();cc++) {
+    for (cc=0; (size_t)cc<sym.entries.size();cc++) {
         if ((sym.entries[cc].sscope > from_level) && (sym.entries[cc].stype == SYM_LOCALVAR)) {
             // caller will sort out stack, so ignore parameters
             if ((sym.entries[cc].flags & SFLG_PARAMETER)==0) {
